Move out convert_t5_weights to its own file
