+++ transformer_lens/pretrained/weight_conversions/__init__.py
@@ -16,4 +16,5 @@ from .phi3 import convert_phi3_weights
 from .gemma import convert_gemma_weights
 from .mingpt import convert_mingpt_weights
 from .nanogpt import convert_nanogpt_weights
+from .t5 import convert_t5_weights
 from .neel_solu_old import convert_neel_solu_old_weights
